src/tb/mac_pcs/requirements.txt
cocotb==1.7.2
cocotb-bus==0.2.1
cocotb-test==0.2.4
cocotbext-axi==0.1.18
debugpy==1.6.3
numpy==1.23.5
pytest==7.1.3
pyuvm==2.9.0
PyYAML==6.0
